Added function tracing namespace using `q` module.
Simply do `from toolz.traced import *` and viola!  In another termal, watch
the output real-time via `tail -f /tmp/toolz`.

To trace a user function use `trace` as a decorator or function.

This is meant as a proof-of-concept.  `q` was copied from
https://github.com/zestyping/q
and was slightly modified to output to "/tmp/toolz" instead of "/tmp/q".