`nth` accepts a list of indices to get just like `get`
See #97.  A list of indices also works on iterators.

`nth` also now works correctly on mapping types.  This was achieved by
checking for the attribute `__add__`, because the sequence protocol
includes concatenation (via addition), but mapping and iterator protocols
do not.  This is true in python2 and python3.  Checking
`isinstance(seq, collections.Sequence)` would be safer, but it is
significantly slower.  An alternative way to test if `seq` is a sequence
is to try to perform a slice operation on it.